+++ ontolearn/learners/__init__.py
@@ -1,0 +1,1 @@
+from .drill import Drill
