+++ ontolearn/learners/nero.py
@@ -1,0 +1,9 @@
+class NERO:
+    def __init__(self):
+        pass
+
+    def train(self):
+        pass
+
+    def fit(self):
+        pass
